+++ public/index.html
@@ -6,10 +6,10 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
-      name="description"
-      content="Web site created using create-react-app"
+      name="Phoenix Robotics Website"
+      content="Official website for the Phoenix Robotics student club at the TUM"
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logos/phoenix.png" />
+    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logos/phoenix.svg" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
