+++ src/App.css
@@ -14,6 +14,14 @@ body {
   font-weight: 500;
 }
 
+img {
+  max-width: 100%;
+}
+
+a {
+  text-decoration: none;
+}
+
 h1 {
   font-size: 64px;
   font-weight: 600;
