+++ src/pages/home/style.css
@@ -1,12 +1,16 @@
+.wide {
+  width: 100%;
+  margin-bottom: -10px;
+}
+
 .center {
   margin-left: auto;
   margin-right: auto;
   width: 70%;
 }
 
-.wide {
-  width: 100%;
-  margin-bottom: -10px;
+.center img {
+  max-height: 40vh;
 }
 
 .logo {
@@ -26,10 +30,6 @@
   color: var(--beige);
   background-color: var(--dark-blue);
   display: flow-root;
-}
-
-.clear {
-  clear: both;
 }
 
 .float_right {
